fix(security): address audit follow-ups for CI and Collection exports
Run the full CI workflow on every push and PR (remove paths-ignore) so
markdown-only changes cannot skip checks. Move Collection summary, table,
and date helpers into collectionPresentation.jsx so the route module only
default-exports the page; tests import presentation pieces explicitly.
Document why App.test.jsx mocks Firebase and point to App.hooks.test.jsx
for auth coverage.
